requirements.txt
accelerate
datasets
faiss-cpu
huggingface_hub
numpy
pyarrow
safetensors
torch
tqdm
transformers
